requirements.txt
#pip install -r requirements.txt

mailchimp3==3.0.6
flask_cors==3.0.10
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
mysql-connector-python==8.0.27
requests==2.27.1
PyMySQL==1.0.2
mailchimp-marketing==3.0.74
urllib3==1.26.9
PyMySQL==1.0.2
PyJWT==2.3.0
stripe==2.69.0
